Cambios para soportar mejor numeros grandes en las compras pruebas realizadas con montos de 21 millones

+++ src/manejo_productos.c
@@ -32,15 +32,15 @@ CreateNew (gchar *barcode, gdouble cantidad)
 {
   Productos *new = NULL;
   //  PGresult *res = EjecutarSQL (g_strdup_printf ("SELECT productos* FROM productos WHERE barcode='%s'", barcode));
-  PGresult *res = EjecutarSQL 
-    (g_strdup_printf 
+  PGresult *res = EjecutarSQL
+    (g_strdup_printf
      ("SELECT codigo, barcode, descripcion, marca, contenido, unidad, precio, fifo, margen_promedio, (SELECT monto FROM impuestos WHERE id=0 AND "
       "productos.impuestos='t'), (SELECT monto FROM impuestos WHERE id=productos.otros), canje , stock_pro, precio_mayor, cantidad_mayor, mayorista FROM productos WHERE barcode='%s'", barcode));
- 
+
   new = (Productos *) g_malloc (sizeof (Productos));
-  
+
   new->product = (Producto *) g_malloc (sizeof (Producto));
-  
+
   new->product->cantidad = cantidad;
   new->product->codigo = PQgetvalue (res, 0, 0);
   new->product->barcode = PQgetvalue (res, 0, 1);
@@ -69,15 +69,15 @@ FreeProduct (Productos *productos)
 {
   /*
   Producto *producto = productos->product;
-  
-  free (producto->codigo);  
-   
+
+  free (producto->codigo);
+
   g_free (producto->producto);
   g_free (producto->marca);
   g_free (producto->unidad);
   g_free (producto->barcode);
   g_free (producto);
-  
+
   g_free (productos->product);
   g_free (productos);
   */
@@ -87,7 +87,7 @@ gint
 AgregarALista (gchar *codigo, gchar *barcode, gdouble cantidad)
 {
   Productos *new = NULL;
-      
+
   if (venta->header == NULL)
     {
       new = CreateNew (barcode, cantidad);
@@ -97,7 +97,7 @@ AgregarALista (gchar *codigo, gchar *barcode, gdouble cantidad)
 
       new->back = NULL;
       new->next = venta->header;
-      
+
       new->product->lugar = 1;
     }
   else
@@ -105,18 +105,18 @@ AgregarALista (gchar *codigo, gchar *barcode, gdouble cantidad)
       Productos *end = venta->header;
 
       new = CreateNew (barcode, cantidad);
-      
+
       while (end->next != venta->header)
 	end = end->next;
-      
+
       new->back = end;
- 
+
       end->next = new;
-      
+
       new->next = venta->header;
-      
+
       venta->products = new;
-      
+
       new->product->lugar = new->back->product->lugar+1;
     }
 
@@ -129,7 +129,7 @@ EliminarDeLista (gchar *codigo, gint position)
   Productos *find = venta->header;
   Productos *end = venta->header;
 
-  /* 
+  /*
      Must be re-write it, to fix and prevent any kind of error
   */
 
@@ -142,7 +142,7 @@ EliminarDeLista (gchar *codigo, gint position)
   if (find == venta->header && venta->header->next == venta->header)
     {
       FreeProduct (find);
-      
+
       venta->header = NULL;
       venta->products = NULL;
     }
@@ -157,7 +157,7 @@ EliminarDeLista (gchar *codigo, gint position)
   else if (find == end)
     {
       find->back->next = venta->header;
-      
+
       FreeProduct (find);
     }
   else
@@ -190,11 +190,11 @@ CalcularTotal (Productos *header)
 	total += (gdouble)(cal->product->precio_mayor * cal->product->cantidad);
       else
 	total += (gdouble)(cal->product->precio * cal->product->cantidad);
-      
+
       cal = cal->next;
     }
   while (cal != header);
-  
+
   return total;
 }
 
@@ -204,15 +204,15 @@ ReturnTotalProducts (Productos *header)
   Productos *cal = header;
   gint total = 0;
 
-  do 
+  do
     {
       total += 1;
-      
+
       cal = cal->next;
     }
   while (cal != header);
 
-  return total;  
+  return total;
 }
 
 /*gchar *
@@ -221,7 +221,7 @@ ReturnAllProductsCode (Productos *header)
   Productos *cal = header;
   gchar *codes = (gchar *) g_malloc0 (600);
   gchar *alter = (gchar *) g_malloc0 (600);
-  
+
   do
     {
       if (strcmp (codes, "") == 0)
@@ -230,11 +230,11 @@ ReturnAllProductsCode (Productos *header)
 	  g_snprintf (alter, 600, "%s", codes);
 	}
       else
-	{	  
+	{
 	  g_snprintf (codes, 600, "%s%s_%d ", alter, cal->product->codigo, cal->product->cantidad);
 	  g_snprintf (alter, 600, "%s", codes);
 	}
-	
+
 
       cal = cal->next;
     }
@@ -251,14 +251,14 @@ ListClean (void)
   gint total = ReturnTotalProducts (venta->header);
   gint i;
 
-  
+
   for (i = 0; i < total; i++)
     {
       tofree = alter;
       alter = tofree->next;
-      FreeProduct (tofree);      
-    }  
-  
+      FreeProduct (tofree);
+    }
+
   venta->header = NULL;
   venta->products = NULL;
 
@@ -278,16 +278,16 @@ CompraListClean (void)
     {
       tofree = alter;
       alter = tofree->next;
-      FreeProduct (tofree);      
+      FreeProduct (tofree);
     }
   */
 
   do {
-    
+
     tofree = alter;
-    
-    FreeProduct (tofree);    
-    
+
+    FreeProduct (tofree);
+
     tofree = alter->next;
 
     g_free (alter);
@@ -300,14 +300,14 @@ CompraListClean (void)
   compra->products_list = NULL;
   compra->current = NULL;
 
-  return 0; 
+  return 0;
 }
 
 Productos *
 CompraCreateNew (gchar *barcode, double cantidad, gint precio_final, gdouble precio_compra, gint margen)
 {
   Productos *new = NULL;
-  PGresult *res = EjecutarSQL 
+  PGresult *res = EjecutarSQL
     (g_strdup_printf ("SELECT codigo, barcode, descripcion, marca, contenido, unidad, perecibles, canje, stock_pro, tasa_canje, precio_mayor, cantidad_mayor, mayorista FROM productos "
 		      "WHERE barcode='%s'", barcode));
 
@@ -323,7 +323,7 @@ CompraCreateNew (gchar *barcode, double cantidad, gint precio_final, gdouble pre
   new->product->contenido = atoi (PQgetvalue (res, 0, 4));
   new->product->unidad = PQgetvalue (res, 0, 5);
   //  new->product->precio = atoi (PQgetvalue (res, 0, 8));
-  new->product->margen = margen;  
+  new->product->margen = margen;
   new->product->iva = GetIVA (barcode);
   new->product->otros = GetOtros (barcode);
   new->product->precio = precio_final;
@@ -345,38 +345,38 @@ CompraCreateNew (gchar *barcode, double cantidad, gint precio_final, gdouble pre
 }
 
 gint
-CompraAgregarALista (gchar *barcode, gdouble cantidad, gint precio_final, gdouble precio_compra, 
+CompraAgregarALista (gchar *barcode, gdouble cantidad, gint precio_final, gdouble precio_compra,
 		     gint margen, gboolean ingreso)
 {
   Productos *new = NULL;
-  
+
   if (ingreso == TRUE)
     {
       if (compra->header == NULL)
 	{
 	  new = CompraCreateNew (barcode, cantidad, precio_final, precio_compra, margen);
 	  compra->header = new;
-	  
+
 	  compra->products_list = new;
-	  
+
 	  new->back = NULL;
 	  new->next = compra->header;
 	}
       else
-	{    
+	{
 	  Productos *end = compra->header;
 
 	  new = CompraCreateNew (barcode, cantidad, precio_final, precio_compra, margen);
-	  
+
 	  while (end->next != compra->header)
 	    end = end->next;
-	  
+
 	  new->back = end;
-	
+
 	  end->next = new;
-	  
+
 	  new->next = compra->header;
-	  
+
 	  compra->products_list = new;
 	}
     }
@@ -386,9 +386,9 @@ CompraAgregarALista (gchar *barcode, gdouble cantidad, gint precio_final, gdoubl
 	{
 	  new = CompraCreateNew (barcode, cantidad, precio_final, precio_compra, margen);
 	  compra->header_compra = new;
-	  
+
 	  compra->products_compra = new;
-	  
+
 	  new->back = NULL;
 	  new->next = compra->header_compra;
 	  compra->current = new->product;
@@ -398,16 +398,16 @@ CompraAgregarALista (gchar *barcode, gdouble cantidad, gint precio_final, gdoubl
 	  Productos *end = compra->header_compra;
 
 	  new = CompraCreateNew (barcode, cantidad, precio_final, precio_compra, margen);
-	  
+
 	  while (end->next != compra->header_compra)
 	    end = end->next;
-	  
+
 	  new->back = end;
-	  
+
 	  end->next = new;
-	  
+
 	  new->next = compra->header_compra;
-	  
+
 	  compra->products_compra = new;
 	  compra->current = new->product;
 	}
@@ -431,7 +431,7 @@ DropBuyProduct (gchar *codigo)
   if (find == compra->header_compra && compra->header_compra->next == compra->header_compra)
     {
       FreeProduct (find);
-      
+
       compra->header_compra = NULL;
       compra->products_compra = NULL;
     }
@@ -445,7 +445,7 @@ DropBuyProduct (gchar *codigo)
   else if (find == end)
     {
       find->back->next = compra->header_compra;
-      
+
       FreeProduct (find);
     }
   else
@@ -475,14 +475,14 @@ SearchProductByBarcode (gchar *barcode, gboolean ingreso)
       find = compra->header_compra;
       control = compra->header_compra;
     }
-  
-  do 
+
+  do
     {
       if (strcmp (find->product->barcode, barcode) == 0)
 	return find->product;
       else
 	find = find->next;
-    } 
+    }
   while (find != control);
 
   return NULL;
@@ -518,11 +518,11 @@ gdouble
 CalcularTotalCompra (Productos *header)
 {
   Productos *cal = header;
-  gdouble total = 0;
-  
+  guint64 total = 0;
+
   if (cal == NULL)
     return total;
-  
+
   do
     {
       //      if (cal->product->precio_compra == 0)
@@ -533,7 +533,7 @@ CalcularTotalCompra (Productos *header)
       cal = cal->next;
     }
   while (cal != header);
-  
+
   return total;
 }
